Add optional Java (class)  migration support
A flyway task may now be run after the compilation phase and the class
migrations will be included. This is an optional configuration as some
build processes may instead want to run migrations prior to the
compilation task(s), e.g. to perform schema code generation. [1]

The resources are no longer a manditory location for SQL migrations.
This is now a nice default setting if the locations have not been
configured. This supports configurations where SQL scripts are
database-specific and separated by a folders. [2]

[1] http://flywaydb.org/documentation/migration/java.html
[2] http://flywaydb.org/documentation/faq.html#db-specific-sql